Maajor changes in code structure
