+++ configs/DeepQLearning.cfg
@@ -1,0 +1,22 @@
+[simulation_params]
+load_data:                  False
+load_data_path:             DeepNet/models/Tello_indoor/VanLeer/
+
+[RL_params]
+input_size:                 227
+num_actions:                25
+train_type:                 e2e
+wait_before_train:          10000
+max_iters:                  150000
+buffer_len:                 30000
+batch_size:                 32
+epsilon_saturation:         150000
+crash_thresh:               1.3
+Q_clip:                     True
+train_interval:             3
+update_target_interval:     80000
+gamma:                      0.99
+dropout_rate:               0.1
+learning_rate:              1e-6
+switch_env_steps:           2000
+epsilon_model:              exponential
