+++ configs/config.cfg
@@ -8,33 +8,10 @@ phase:                      train
 SimMode:                    Multirotor
 drone:                      DJIMavic
 ClockSpeed:                 3
+algorithm:                  DeepQLearning
+ip_address:                 127.0.0.2
 
 [camera_params]
 width:                      640
 height:                     360
 fov_degrees:                80
-
-
-[simulation_params]
-ip_address:                 127.0.0.2
-load_data:                  False
-load_data_path:             DeepNet/models/Tello_indoor/VanLeer/
-
-[RL_params]
-input_size:                 227
-num_actions:                25
-train_type:                 e2e
-wait_before_train:          10000
-max_iters:                  150000
-buffer_len:                 30000
-batch_size:                 32
-epsilon_saturation:         150000
-crash_thresh:               1.3
-Q_clip:                     True
-train_interval:             3
-update_target_interval:     80000
-gamma:                      0.99
-dropout_rate:               0.1
-learning_rate:              1e-6
-switch_env_steps:           2000
-epsilon_model:              exponential
